feat: add ability to requests changes based on document from site
By clicking the edit icon at the top right of a page a user will either
be redirected to a github edit and commit landing page for the specific
page it was clicked in or be directed to the fork repo process and
facilitate the editing and committing workflow from there.

Adds small icon next to the print and github icons.