refactor: Switch to perl regex for non-greedy quantifiers
The initial sed to extract markdown links from lines was flawed in that
it would pull all text up to the ending parenthese, which was causing it
to grab unwanted text on lines that contained markdown links inside
parentheses.
I switched to perl to use non-greedy quantifiers to prevent this.

I also added a check for the unitindex.md file before attempting to pull
the unit numbers and titles from it.